refactor(plugin-ts,plugin-zod): share buildOptionsSchema via internals/shared
plugin-ts's buildOptions and plugin-zod's buildOptionsSchema built the exact
same { body, path, query, headers } object schema from the same inputs,
differing only in a primitive: 'object' marker that's a no-op for the TS
printer. Consolidates both into internals/shared's buildOptionsSchema,
typed against the existing OperationTypeNameResolver shape both plugins'
resolvers already satisfy. No generated output changes (verified via each
package's test suite and the repo-root integration suite).
