+++ internals/shared/src/index.ts
@@ -1,6 +1,7 @@
 export {
   buildClientOptionType,
   buildOperationComments,
+  buildOptionsSchema,
   buildRequestConfigType,
   buildRequestParamsSignature,
   findSuccessStatusCode,
